// src/permission/subject.enum.ts
export enum SubjectsEnum {
  ALL = 'all',
  PERMISSION = 'permission',
  ROLE = 'role',
  USER = 'USER',
  MENU1 = 'menu1',
  MENU11 = 'menu11',
  MENU12 = 'menu12',
  MENU121 = 'menu121',
  MENU122 = 'menu122',
  MENU13 = 'menu13',
  MENU131 = 'menu131',
  MENU2 = 'menu2',
  MENU21 = 'menu21',
  MENU22 = 'menu22',
  MENU221 = 'menu221',
  MENU222 = 'menu222',
  MENU2221 = 'menu2221',
  MENU2222 = 'menu2222',
  MENU223 = 'menu223',
  MENU23 = 'menu23',
  MENU3 = 'menu3',
  MENU31 = 'menu31',
  MENU32 = 'menu32',
}